refactor: clean cluster/permissions endpoint
