+++ gravitee-apim-console-webui/src/services-ngx/cluster-member.service.ts
@@ -20,9 +20,7 @@ import { Inject, Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 
 import { Constants } from '../entities/Constants';
-import {
-  TransferOwnership
-} from '../management/clusters/details/user-permissions/transfer-ownership/cluster-transfer-ownership.component';
+import { TransferOwnership } from '../management/clusters/details/user-permissions/transfer-ownership/cluster-transfer-ownership.component';
 
 @Injectable({
   providedIn: 'root',
@@ -32,10 +30,6 @@ export class ClusterMemberService {
     private readonly http: HttpClient,
     @Inject(Constants) private readonly constants: Constants,
   ) {}
-
-  getPermissions(clusterId: string): Observable<Record<string, string>> {
-    return this.http.get<Record<string, string>>(`${this.constants.env.v2BaseURL}/clusters/${clusterId}/members/permissions`);
-  }
 
   getMembers(clusterId: string, page = 1, perPage = 10): Observable<MembersResponse> {
     return this.http.get<MembersResponse>(`${this.constants.env.v2BaseURL}/clusters/${clusterId}/members`, {
